requirements.txt
anyio==3.6.2
bcrypt==4.0.1
cffi==1.15.1
cryptography==38.0.3
deta==1.1.0
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.86.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.1
nanoid==2.0.0
passlib==1.7.4
Pillow==9.3.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
python-jose==3.3.0
python-multipart==0.0.5
qrcode==7.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
